Functioning L2 LP refining

+++ include/mtkahypartypes.h
@@ -81,8 +81,8 @@ typedef struct {
 
 typedef unsigned long int mt_kahypar_hypernode_id_t;
 typedef unsigned long int mt_kahypar_hyperedge_id_t;
-typedef float mt_kahypar_hypernode_weight_t;
-typedef float mt_kahypar_hyperedge_weight_t;
+typedef int mt_kahypar_hypernode_weight_t;
+typedef int mt_kahypar_hyperedge_weight_t;
 typedef int mt_kahypar_partition_id_t;
 
 /**
